requirements_test.txt
pytest-asyncio>=0.23.6
pytest-cov>=5.0.0
pytest>=8.1.1
python-dotenv>=1.0.1
aioresponses>=0.7.6